chore: move PrintableProcedure to procedures file, add iterable getter
